# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditAction": {
        "enum": [
          "CREATE",
          "UPDATE",
          "DELETE",
          "LOGIN",
          "LOGOUT"
        ],
        "title": "AuditAction",
        "type": "string"
      },
      "AuditEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AuditAction"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "changes": {
            "additionalProperties": true,
            "title": "Changes",
            "type": "object"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "outcome": {
            "$ref": "#/components/schemas/AuditOutcome"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "audit_id",
          "actor_id",
          "action",
          "resource_type",
          "resource_id",
          "timestamp",
          "ip_address",
          "outcome"
        ],
        "title": "AuditEntryResponse",
        "type": "object"
      },
...
      "AuditOutcome": {
        "enum": [
          "SUCCESS",
          "FAILURE"
        ],
        "title": "AuditOutcome",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "maxLength": 36,
                "minLength": 36,
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "mitre_tactic": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Tactic"
          },
          "mitre_technique": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Technique"
          },
          "severity": {
            "$ref": "#/components/schemas/IncidentSeverity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus",
            "default": "OPEN"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 200,
            "minLength": 5,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "severity"
        ],
        "title": "IncidentCreateRequest",
        "type": "object"
      },
      "IncidentDeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "incident": {
            "$ref": "#/components/schemas/IncidentResponse"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "incident"
        ],
        "title": "IncidentDeleteResponse",
        "type": "object"
      },
      "IncidentListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "IncidentListResponse",
        "type": "object"
      },
      "IncidentResponse": {
        "additionalProperties": false,
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "is_deleted": {
            "title": "Is Deleted",
            "type": "boolean"
          },
          "mitre_tactic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Tactic"
          },
          "mitre_technique": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Technique"
          },
          "severity": {
            "$ref": "#/components/schemas/IncidentSeverity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "title",
          "description",
          "severity",
          "status",
          "created_by",
          "assigned_to",
          "mitre_tactic",
          "mitre_technique",
          "tags",
          "is_deleted",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
      "IncidentSeverity": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "IncidentSeverity",
        "type": "string"
      },
      "IncidentStatus": {
        "enum": [
          "OPEN",
          "IN_PROGRESS",
          "CONTAINED",
          "RESOLVED",
          "CLOSED"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentTimelineResponse": {
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "incident_id",
          "items",
          "total"
        ],
        "title": "IncidentTimelineResponse",
        "type": "object"
      },
      "IncidentUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "maxLength": 36,
                "minLength": 36,
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "mitre_tactic": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Tactic"
          },
          "mitre_technique": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Technique"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 10,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 5,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "IncidentUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Secure Incident Management API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/incidents/": {
      "get": {
        "operationId": "list_incidents_route_incidents__get",
        "parameters": [
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "assigned_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "in": "query",
            "name": "created_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "in": "query",
            "name": "created_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Incidents Route",
        "tags": [
          "incidents"
        ]
      },
      "post": {
        "operationId": "create_incident_route_incidents__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Incident Route",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}": {
      "delete": {
        "operationId": "delete_incident_route_incidents__incident_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Incident Route",
        "tags": [
          "incidents"
        ]
      },
      "get": {
        "operationId": "get_incident_route_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Incident Route",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "update_incident_route_incidents__incident_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Incident Route",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/incidents/{incident_id}/timeline": {
      "get": {
        "operationId": "get_incident_timeline_route_incidents__incident_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentTimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Incident Timeline Route",
        "tags": [
          "incidents"
        ]
      }
    }
  }
}